Feat: seperate pathMap for resue
