+++ src/constants/coffee.ts
@@ -12,3 +12,16 @@ export const COFFEE_TEXTS = {
   including: '포함된 카페인 함량은 ',
   is: '입니다.'
 };
+
+export const BRANDS = [
+  '스타벅스',
+  '이디야',
+  '메가커피',
+  '할리스',
+  '파스쿠찌',
+  '엔젤리너스',
+  '더벤티',
+  '빽다방',
+  '컴포즈커피',
+  '폴바셋'
+];
